requirements.txt
pandas
wordcloud
arabic-reshaper
python-bidi
openpyxl